added snapshot test for index.js and put together all book snapshot tests
