+++ src/__tests__/index.snapshot.js
@@ -1,0 +1,5 @@
+import Index from '../index.js';
+
+it('renders without crashing', () => {
+  expect(JSON.stringify(Index)).toMatchSnapshot();
+});
